+++ stable_baselines3/a2c/a2c.py
@@ -120,7 +120,7 @@ class A2C(OnPolicyAlgorithm):
         rollout buffer (one gradient step over whole data).
         """
         # Switch to train mode (this affects batch norm / dropout)
-        self.policy.train()
+        self.policy.set_training_mode(True)
 
         # Update optimizer learning rate
         self._update_learning_rate(self.policy.optimizer)